Fix C++11 compatibility

+++ Development/nmos/control_protocol_utils.cpp
@@ -502,7 +502,7 @@ namespace nmos
                         // only update pending received time if not already set
                         if (pending_received_time.as_integer() == 0)
                         {
-                            if (!set_property(resources, oid, status_pending_received_time_field_name, now_time, gate))
+                            if (!set_property(resources, oid, status_pending_received_time_field_name, static_cast<int64_t>(now_time), gate))
                             {
                                 return false;
                             }
